Fixed flags compilation errors

+++ mandatory/src/map_check3.c
@@ -6,7 +6,7 @@
 /*   By: zouddach <zouddach@1337.student.ma>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/08 12:10:06 by zouddach          #+#    #+#             */
-/*   Updated: 2025/02/08 13:00:11 by zouddach         ###   ########.fr       */
+/*   Updated: 2025/02/09 19:23:56 by zouddach         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -130,7 +130,7 @@ bool	check_wall_collision(t_game *game, int keycode)
 		&& check_right_collision(game, &new_x, &new_y) == false)
 		return (false);
 	if (new_x < 0 || new_y < 0 || new_x >= game->map.rows
-		|| new_y >= ft_strlen(game->map.map[(int)new_x]))
+		|| (size_t)new_y >= ft_strlen(game->map.map[(int)new_x]))
 		return (false);
 	if (game->map.map[(int)new_x][(int)new_y] == '1')
 		return (false);
